Move hover unittests to separate file
