+++ test/test_server.py
@@ -381,138 +381,3 @@ def test_refs():
             [free_path, 78, 6, 12],
         ),
     )
-
-
-def test_hover():
-    def hover_req(file_path: str, ln: int, col: int) -> str:
-        return write_rpc_request(
-            1,
-            "textDocument/hover",
-            {
-                "textDocument": {"uri": str(file_path)},
-                "position": {"line": ln, "character": col},
-            },
-        )
-
-    def check_return(result_array, checks):
-        assert len(result_array) == len(checks)
-        for (i, check) in enumerate(checks):
-            assert result_array[i]["contents"][0]["value"] == check
-
-    #
-    string = write_rpc_request(1, "initialize", {"rootPath": str(test_dir)})
-    file_path = test_dir / "subdir" / "test_abstract.f90"
-    string += hover_req(file_path, 7, 30)
-    file_path = test_dir / "hover" / "parameters.f90"
-    string += hover_req(file_path, 2, 28)
-    string += hover_req(file_path, 3, 28)
-    string += hover_req(file_path, 4, 28)
-    string += hover_req(file_path, 4, 41)
-    string += hover_req(file_path, 6, 28)
-    string += hover_req(file_path, 7, 38)
-    string += hover_req(file_path, 7, 55)
-    string += hover_req(file_path, 8, 37)
-    string += hover_req(file_path, 8, 50)
-    string += hover_req(file_path, 9, 37)
-    string += hover_req(file_path, 9, 48)
-    string += hover_req(file_path, 10, 37)
-    string += hover_req(file_path, 10, 48)
-    file_path = test_dir / "hover" / "pointers.f90"
-    string += hover_req(file_path, 1, 26)
-    file_path = test_dir / "hover" / "functions.f90"
-    string += hover_req(file_path, 1, 11)
-    string += hover_req(file_path, 7, 19)
-    string += hover_req(file_path, 12, 12)
-    string += hover_req(file_path, 18, 19)
-    string += hover_req(file_path, 23, 34)
-    string += hover_req(file_path, 28, 11)
-    string += hover_req(file_path, 34, 21)
-    string += hover_req(file_path, 46, 11)
-    string += hover_req(file_path, 51, 11)
-    string += hover_req(file_path, 55, 11)
-    file_path = test_dir / "hover" / "spaced_keywords.f90"
-    string += hover_req(file_path, 1, 45)
-    string += hover_req(file_path, 2, 99)
-    file_path = test_dir / "hover" / "recursive.f90"
-    string += hover_req(file_path, 9, 40)
-    file_path = test_dir / "subdir" / "test_submod.F90"
-    string += hover_req(file_path, 29, 24)
-    string += hover_req(file_path, 34, 24)
-    file_path = test_dir / "test_diagnostic_int.f90"
-    string += hover_req(file_path, 19, 14)
-
-    errcode, results = run_request(string, fortls_args=["--sort_keywords"])
-    assert errcode == 0
-    #
-    ref_results = (
-        """SUBROUTINE test(a, b)
- INTEGER(4), DIMENSION(3,6), INTENT(IN) :: a
- REAL(8), DIMENSION(4), INTENT(OUT) :: b""",
-        "INTEGER, PARAMETER :: var = 1000",
-        "INTEGER",
-        "INTEGER, PARAMETER :: var2 = 23",
-        "INTEGER, PARAMETER :: var3 = var*var2",
-        "INTEGER, PARAMETER :: var4 = 123",
-        "DOUBLE PRECISION, PARAMETER :: somevar = 23.12",
-        "DOUBLE PRECISION, PARAMETER :: some = 1e-19",
-        "LOGICAL(kind=8), PARAMETER :: long_bool = .true.",
-        "LOGICAL",
-        "CHARACTER(len=5), PARAMETER :: sq_str = '12345'",
-        "CHARACTER(LEN=5)",
-        'CHARACTER(len=5), PARAMETER :: dq_str = "12345"',
-        "CHARACTER(LEN=5)",
-        "INTEGER, POINTER",
-        """FUNCTION fun1(arg) RESULT(fun1)
- INTEGER, INTENT(IN) :: arg
- INTEGER :: fun1""",
-        """FUNCTION fun2(arg) RESULT(fun2)
- INTEGER, INTENT(IN) :: arg
- INTEGER :: fun2""",
-        """FUNCTION fun3(arg) RESULT(retval)
- INTEGER, INTENT(IN) :: arg
- INTEGER :: retval""",
-        """FUNCTION fun4(arg) RESULT(retval)
- INTEGER, INTENT(IN) :: arg
- INTEGER :: retval""",
-        # Notice that the order of the modifiers does not match the source code
-        # This is part of the test, ideally they would be identical but previously
-        # any modifiers before the type would be discarded
-        """PURE ELEMENTAL FUNCTION fun5(arg) RESULT(retval)
- INTEGER, INTENT(IN) :: arg
- INTEGER :: retval""",
-        """FUNCTION fun6(arg) RESULT(retval)
- INTEGER, INTENT(IN) :: arg
- INTEGER, DIMENSION(10,10) :: retval""",
-        """PURE FUNCTION outer_product(x, y) RESULT(outer_product)
- REAL, DIMENSION(:), INTENT(IN) :: x
- REAL, DIMENSION(:), INTENT(IN) :: y
- REAL, DIMENSION(SIZE(X), SIZE(Y)) :: outer_product""",
-        """FUNCTION dlamch(cmach) RESULT(dlamch)
- CHARACTER :: CMACH""",
-        """FUNCTION fun7() RESULT(val)
- TYPE(c_ptr) :: val""",
-        """TYPE(c_ptr) FUNCTION c_loc(x) RESULT(c_loc)""",
-        """REAL, DIMENSION(:, :), INTENT(IN)""",
-        """REAL, DIMENSION( SIZE(ARG1, 1), MAXVAL([SIZE(ARG1, 2), """
-        """SIZE(ARG1, 1)]) ), INTENT(OUT)""",
-        """RECURSIVE SUBROUTINE recursive_assign_descending(node, vector, current_loc)
- TYPE(tree_inode), POINTER, INTENT(IN) :: node
- INTEGER, DIMENSION(:), INTENT(INOUT) :: vector
- INTEGER, INTENT(INOUT) :: current_loc""",
-        """FUNCTION point_dist(a, b) RESULT(distance)
- TYPE(point), INTENT(IN) :: a
- TYPE(point), INTENT(IN) :: b
- REAL :: distance""",
-        """FUNCTION is_point_equal_a(a, b) RESULT(is_point_equal_a)
- TYPE(point), INTENT(IN) :: a
- TYPE(point), INTENT(IN) :: b
- LOGICAL :: is_point_equal_a""",
-        # Could be subject to change
-        """FUNCTION foo2(f, g, h) RESULT(arg3)
- FUNCTION f(x) :: f
- FUNCTION g(x) :: g
- FUNCTION h(x) :: h
- REAL :: arg3""",
-    )
-    assert len(ref_results) == len(results) - 1
-    check_return(results[1:], ref_results)
